[BEAM-7126] Remove LengthPrefixCoder from Flink's key Coder
Unknown Coders are wrapped inside a LengthPrefixCoder to ship serialized data
without knowledge of the Coder used to serialize the data. When operators use
state or timers, we rely on data to be partitioned by key. We included the
LengthPrefixCoder for the key Coder which led to issues with handling state
requests which do not include the length prefix in the key encoding.

We can safely remove the LengthPrefixCoder from the key because we do not store
the key in a nested context where we might need the length prefix.

This also expands the state cleanup tests to further verify that state is
cleaned up correctly.