[BEAM-5428][backport] Modify cache token Proto design to only include tokens in ProcessBundleRequest
On the mailing list, we discussed simplifying the current cache token design to
allow for a simpler asynchronous implementation on the SDK side. The old design
transferred valid cache tokens in ProcessBundleRequests which were previously
assigned in get/put/clear state requests.

The new proposal drops the cache token assignment in get/put/clear requests in
favor of a predefined cache token list for all user state and each side
input. These cache tokens are valid for the entire bundle and do not change. The
Runner needs to take care to generate new user state cache tokens in case of a
failure and to properly update the side input cache tokens when the side input
gets updated.