[BEAM-8157] Ensure key encoding for state requests is consistent across SDKs
The key encoding for state requests across the SDKs, i.e. Python and Java, is
not consistent. Java uses a NESTED encoding, whereas Python uses OUTER. We
should settle on one encoding.

The Flink runner so far uses OUTER encoding, but it seems sensible to switch to
NESTED encoding since inferring the OUTER encoding for a key coder in a KV<K,V>
structure is a potential source of error.