Checkpointing fixes part of Beam 2.17.0 (#29)
* [BEAM-8549] Do not use keyed operator state for checkpoint buffering

The current buffer logic for items emitted during checkpointing is faulty in the
sense that the buffer is partitioned on the output keys of the operator. The key
may be changed or even be dropped. Thus, the original key partitioning will not
be maintained which will cause checkpointing to fail.

An alternative solution would be BEAM-6733 / #9652, but this change keeps the
current buffering logic in place. The output buffer may now always be
redistributed round-robin upon restoring from a checkpoint. Note that this is
fine because no assumption can be made about the distribution of output elements
of a DoFn operation.

* [BEAM-8566] Fix checkpoint buffering when when another bundle is started during checkpointing

As part of a checkpoint, the current bundle is finalized. When the bundle is
finalized, the watermark, currently held back, may also be progressed which can
cause the start of another bundle. When a new bundle is started, any
to-be-buffered items from the previous bundle for the pending checkpoint may be
emitted. This should not happen.

This only effects portable pipelines where we have to hold back the watermark
due to the asynchronous processing of elements.

* [BEAM-8566] Do not swallow execution errors during checkpointing

If a bundle fails to finalize before creating a checkpoint, it may be swallowed
and just considered a checkpointing error. This breaks the execution flow and
exactly-once guarantees.