[BEAM-8962] Add option to disable the metric container accumulator
The FlinkMetricContainer wraps the Beam metric container for reporting metrics,
but also stores them as Flink accumulators. With high parallelism jobs with over
a thousand tasks and many built-in Beam metrics for every Beam step, this can
accumulate to over 100MB of serialized data which is stored in the JobManager's
ExecutionGraph. This then fails to even sent over the wire, due to the
akka.framesize limit (10MB by default), and manifests in {{500 Internal Server
Error}}s in the web frontend.

This introduces an option to disable the reporting via accumulators. The
accumulator is mostly useful for batch workloads where you can retrieve the
final accumulator values at the end of the job. It adds a lot of memory and
network overhead.

In the future, we might disable it by default on streaming workloads.