Add option to disable the metric container accumulator
This was observed via `500 Internal Server` errors when viewing jobs in the web
frontend. The Beam metric container does not only report metrics but also stores
them as Flink accumulators. With over a thousand tasks and many built-in Beam
metrics, this can accumulate to over 100MB of serialized data which is stored in
the JobManager's ExecutionGraph. This then fails to even sent over the wire due
to the `akka.framesize` limit. This introduces an option to disable the
reporting via accumulators. We don't need it and it adds a lot of runtime
overhead.