[BEAM-9060] Restore stdout/stderr in case Flink's OptimizedPlanEnvironment is used
To extract the JobGraph, Flink calls the main method of a JAR and suppresses the
stdout/stderr. It only prints the output on errors, or when no JobGraph was
found. This is often not desired for debugging purposes and logging purposes.

This change restores the original stdout/stderr when we detect that we are
running inside the `OptimizedPlanEnvironment`.