[BEAM-9161] Ensure non-volatile access of field variables by processing threads
In the DoFnOperator and the subclassed ExecutableStageDoFnOperator there are
effectively two processing threads:

(1) The main processing thread for processing elements and watermarks
(2) A timer thread to support ending bundles after a timeout to optimize for
latency

Although the code was written with a different assumption in the past, the two
threads do not interleave with each other. Only one is active at a time. This is
ensured by Flink's "checkpointLock" which is acquired for every method called on
the operator like processElement, processWatermark, snapshotState. It is also
acquired when timers are fired which are set using Flink's TimeService, like it
is the case for (2).

We've seen issues with bundles being closed multiple times resulting in
exceptions like "Already closed". Very rarely, we've also seen dead bundle ids
resurecting for which no other explanation could be found than an inconsistent
view of the two thread.

This is why this PR makes the fields volatile which are accessed by both
threads. We do not have to synchronize as explained above. We remove the
AtomicBool for the bundle state because it is not necessary, a volatile bool
suffices.