[BEAM-9490] Guard referencing for environment expiration via a lock
When environment expiration is configured, environments are removed from a cache
map after some amount of time. This cache map may be accessed by multiple
ExecutableStage transforms in parallel. During environment expiration there is a
small time window which allows for an environment to still being used while
another transform removes, dereferences, and closes the environment. This is not
expected behavior.