[BEAM-9733] Always let ImpulseSourceFunction emit a final watermark
The Flink Runner's ImpulseSourceFunction does not emit a final watermark, unless
the `--shutdownSourcesOnFinalWatermark` flag has been specified (the flag is
used in tests to shutdown the pipeline after reading all data). Most pipelines
will be long-running and thus do not specify the flag.

Not sending out the final watermark causes GroupByKey to hold back the data of
event time windows until the pipeline is shut down (the final watermark is
always emitted on pipeline shutdown which is why using the above flag works).

[BEAM-9733] Make up for timers set while processing the bundle

Due to the asynchronous nature of the bundle processing which includes access to
state and timers, the output watermark can only be advanced after the current
bundle has finished. Furthermore, we have to take care to only advance it to the
lowest newly set timer in the finished bundle. Otherwise we would be violating
the watermark contract.

Besides the problem of advancing the output watermark too soon, the current
model was also overly restrictive with handling the input watermark. The input
watermark would be pushed back until the end of the bundle. This is not
necessary as we can advance the input watermark freely during bundle processing
which can trigger timers as soon as they become eligible for firing. Special
care has to be taken to handle cleanup timers which may already fire when we
haven't collected all the timers from the SDK Harness. In case we end up with
new timers for a window for which the cleanup timer scheduled a cleanup at the
end of the bundle, we have to defer the cleanup and reschedule the cleanup
timer. Otherwise, we would lose all state for a timer.