[BEAM-10249] Populate state cache with initial values before appending
When new values are added to a bag state, the cache handler in the Python SDK
does not check whether the underlying state already has values present. Although
the append operation will be sent to the backend, the cache itself will be
corrupted because it returns only the newly appended values for get operations,
not the values present before the append operation.

This is a concern when restoring from a checkpoint but also when items are
evicted from the cache.